fix: sync pinned items between library drawer and album art flip menu
Move pinned playlists/albums state into a shared PinnedItemsContext so
all consumers (PlaylistSelection and AlbumArtQuickSwapBack) read from
one React state instance. Previously each component called usePinnedItems()
independently, giving each its own isolated localStorage-backed state;
the storage event only fires for cross-tab changes, so unpinning in the
library drawer was never reflected in the flip menu.

- Add src/contexts/PinnedItemsContext.tsx with PinnedItemsProvider
- Rewrite usePinnedItems.ts to re-export from the context
- Wrap app in PinnedItemsProvider in App.tsx
- Update usePinnedItems tests to render inside PinnedItemsProvider
